requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Deep learning
torch>=2.0.0
tensorflow>=2.13.0

# Signal processing
neurokit2>=0.2.0
biosppy>=2.0.0
heartpy>=1.2.7

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data handling
h5py>=3.8.0
pickle5>=0.0.12

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
